+++ src/editor/components/groupBy/QueryEditorGroupBySection.tsx
@@ -30,7 +30,7 @@ export const QueryEditorGroupBySection = (
 
     if (props.value.expressions.length === 0) {
       return (
-        <QueryEditorSection label={props.label}>
+        <QueryEditorSection label={props.label} tooltip={props.tooltip}>
           <Button
             variant="secondary"
             onClick={() => {
@@ -47,7 +47,7 @@ export const QueryEditorGroupBySection = (
     }
 
     return (
-      <QueryEditorSection label={props.label}>
+      <QueryEditorSection label={props.label} tooltip={props.tooltip}>
         <div className={styles.container}>
           <QueryEditorRepeater id="group-by" onChange={props.onChange} value={props.value}>
             {(childProps) => {